refactor!: initial cleanup of tokenizers (metaschema-update)
- Removed unused Tokenizer classes: `GenePairTokenizer`, `ProteinAlternateTokenizer`
- Removed `NucleotideCache` class. Moved `base_nucleotides` to `Tokenizer` class
- Move `TokenizerBase` methods to `Tokenizer`
- Instance vars --> class vars
