+++ tests/fixtures/tokenizers.yml
@@ -90,7 +90,6 @@ coding_dna_substitution:
     - token: c.292T>C
     - token: (c.233A>G)
     - token: c.509T>A
-    - token: c.54G>H
     - token: c.54G>N
   should_not_match:
     - token: (c.292T>C
@@ -102,14 +101,14 @@ coding_dna_substitution:
     - token: c.509T>
     - token: c.509>A
     - token: c.T>A
+    - token: c.54G>H
 
-genomic_substition:
+genomic_substitution:
   should_match:
     - token: (g.292T>C)
     - token: g.292T>C
     - token: (g.233A>G)
     - token: g.509T>A
-    - token: g.54G>H
     - token: g.54G>N
   should_not_match:
     - token: (g.292T>C
@@ -121,6 +120,7 @@ genomic_substition:
     - token: g.509T>
     - token: g.509>A
     - token: g.T>A
+    - token: g.54G>H
 
 coding_dna_silent_mutation:
   should_match:
