+++ tests/fixtures/validators.yml
@@ -36,12 +36,15 @@ coding_dna_substitution:
     - query: BRAF (c.1799T>A)
     - query: BRAF c.1799T>A
     - query: BRAF V600E c.1799T>A
-    - query: LRG_199t1:c.54G>H
-    - query: LRG_199p1:c.54G>H
-    - query: LRG_199:c.54G>H
+    - query: LRG_199t1:c.54G>N
+    - query: LRG_199p1:c.54G>N
+    - query: LRG_199:c.54G>N
   should_not_match:
     - query: BRAF c.18000000000000T>A
     - query: NM_004333.4:c.17699T>A
+    - query: LRG_199t1:c.54G>H
+    - query: LRG_199p1:c.54G>H
+    - query: LRG_199:c.54G>H
 
 genomic_substitution:
   should_match:
