+++ tests/tokenizers/test_genomic_insertion.py
@@ -1,7 +1,6 @@
 """A module for testing Genomic Insertion Tokenization."""
 import unittest
 
-from variation.tokenizers.caches import NucleotideCache
 from variation.tokenizers import GenomicInsertion
 from .tokenizer_base import TokenizerBase
 
@@ -11,7 +10,7 @@ class TestGenomicInsertionTokenizer(TokenizerBase, unittest.TestCase):
 
     def tokenizer_instance(self):
         """Return Genomic Insertion instance."""
-        return GenomicInsertion(NucleotideCache())
+        return GenomicInsertion()
 
     def token_type(self):
         """Return genomic insertion token type."""
