+++ tests/tokenizers/test_protein_insertion.py
@@ -2,7 +2,6 @@
 import unittest
 
 from variation.tokenizers import ProteinInsertion
-from variation.tokenizers.caches import NucleotideCache
 from .tokenizer_base import TokenizerBase
 
 
@@ -11,7 +10,7 @@ class TestProteinInsertionTokenizer(TokenizerBase, unittest.TestCase):
 
     def tokenizer_instance(self):
         """Return Protein Insertion instance."""
-        return ProteinInsertion(NucleotideCache())
+        return ProteinInsertion()
 
     def token_type(self):
         """Return protein insertion token type."""
