--- variation/tokenizers/caches/__init__.py
@@ -1,2 +1,0 @@
-"""Tokenizer caches package level import."""
-from .nucleotide_cache import NucleotideCache  # noqa: F401
