+++ variation/tokenizers/delins_base.py
@@ -4,20 +4,14 @@ from typing import Optional, Dict, List
 
 from variation.schemas.token_response_schema import DelIns, TokenMatchType, Token
 from .tokenizer import Tokenizer
-from .caches import NucleotideCache
-from .tokenize_base import TokenizeBase
 
 
 class DelInsBase(Tokenizer):
     """Class for tokenizing DelIns."""
 
-    def __init__(self, nucleotide_cache: NucleotideCache) -> None:
-        """Initialize the DelIns Class.
-
-        :param NucleotideCache nucleotide_cache: Valid nucleotides
-        """
+    def __init__(self) -> None:
+        """Initialize the DelIns Class."""
         self.parts = None
-        self.tokenize_base = TokenizeBase(nucleotide_cache)
 
     def match(self, input_string: str) -> Optional[DelIns]:
         """Return tokens that match the input string."""
@@ -77,14 +71,13 @@ class DelInsBase(Tokenizer):
         coordinate_type = parts[0][:1]
         parts[0] = parts[0][2:]
 
-        positions_deleted = self.tokenize_base.get_positions_deleted(parts)
+        positions_deleted = self.get_positions_deleted(parts)
         if not positions_deleted:
             return
         start_pos_del = positions_deleted[0]
         end_pos_del = positions_deleted[1]
 
-        inserted_sequences = \
-            self.tokenize_base.get_transcript_genomic_inserted_sequence(parts)
+        inserted_sequences = self.get_transcript_genomic_inserted_sequence(parts)
         if not inserted_sequences:
             return
         inserted_sequence1 = inserted_sequences[0]
