+++ variation/tokenizers/protein_insertion.py
@@ -5,21 +5,15 @@ from pydantic.error_wrappers import ValidationError
 
 from variation.schemas.token_response_schema import ProteinInsertionToken, \
     TokenMatchType
-from .caches import NucleotideCache
 from .tokenizer import Tokenizer
-from .tokenize_base import TokenizeBase
 
 
 class ProteinInsertion(Tokenizer):
     """Class for tokenizing Insertions on the protein reference sequence."""
 
-    def __init__(self, nucleotide_cache: NucleotideCache) -> None:
-        """Initialize the Protein Insertion Class.
-
-        :param NucleotideCache nucleotide_cache: Valid nucleotides
-        """
+    def __init__(self) -> None:
+        """Initialize the Protein Insertion Class."""
         self.parts = None
-        self.tokenize_base = TokenizeBase(nucleotide_cache)
 
     def match(self, input_string: str) -> Optional[ProteinInsertionToken]:
         """Return token that match the input string."""
@@ -67,13 +61,13 @@ class ProteinInsertion(Tokenizer):
         if len(parts) != 2:
             return
 
-        range_aa_pos = self.tokenize_base.get_aa_pos_range(parts)
+        range_aa_pos = self.get_aa_pos_range(parts)
         if range_aa_pos:
             self.parts["start_aa_flank"] = range_aa_pos[0]
             self.parts["end_aa_flank"] = range_aa_pos[1]
             self.parts["start_pos_flank"] = range_aa_pos[2]
             self.parts["end_pos_flank"] = range_aa_pos[3]
             self.parts["used_one_letter"] = range_aa_pos[4]
-        self.parts["inserted_sequence"] = \
-            self.tokenize_base.get_protein_inserted_sequence(
-                parts, self.parts["used_one_letter"])
+        self.parts["inserted_sequence"] = self.get_protein_inserted_sequence(
+            parts, self.parts["used_one_letter"]
+        )
