+++ variation/tokenizers/silent_mutation.py
@@ -33,9 +33,9 @@ class SilentMutation(PolypeptideSequenceVariationBase):
                 return None
             p_count = input_string.count("p.")
             if p_count == 1:
-                psub_parts = self.splitter.split(input_string)
+                psub_parts = self.splitter_paren_digits.split(input_string)
         else:
-            psub_parts = self.splitter.split(input_string)
+            psub_parts = self.splitter_paren_digits.split(input_string)
 
         self._get_psub(psub_parts)
