+++ variation/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.7.0-dev4"
+__version__ = "0.7.0-dev5"
